(olmv1) Refractor tests for better readability
Also get rid of "exponential backoff with jitter" usage in
`verifyAPIEndpoints` and use an increased timeout of 2min
for "pollUntilContextDeadline" instead.
